refactor: adjust layout components and code formatting
