+++ app/(dashboard)/page.tsx
