+++ package.json
@@ -13,6 +13,7 @@
 		"@radix-ui/react-dialog": "^1.1.1",
 		"@radix-ui/react-dropdown-menu": "^2.1.1",
 		"@radix-ui/react-icons": "^1.3.0",
+		"@radix-ui/react-label": "^2.1.0",
 		"@radix-ui/react-slot": "^1.1.0",
 		"@t3-oss/env-nextjs": "^0.11.0",
 		"@vercel/analytics": "^1.3.1",
